moved personTable.tsx to table component folder
