+++ src/components/myStudents/MyStudents.tsx
@@ -1,6 +1,6 @@
 import { trpc } from "@/client/lib/trpc";
 import React from "react";
-import PersonTable from "../personTable/PersonTable";
+import PersonTable from "../table/table";
 import { Student, StudentHeadCell } from "../table/table";
 
 const MyStudents = () => {
